Initial cut at updated Khronos stylesheet, `khronos2.css`.
This is loosely based on `khronos.css` from
the Khronos internal gitlab `cross-api/stylesheets` repository, which
itself is a fork of the old
link:http://github.com/asciidoctor/asciidoctor-stylesheet-factory[asciidoctor-stylesheet-factory]
repository.

That repository is no longer used by the asciidoctor project.
They have abandoned the SASS format it used to generate the stylesheet,
now writing plain CSS and simplifying it for deployment.

We follow the same model here, basing `khronos2.css` on
`asciidoctor.css`. It adds changes to be closer to the old `colony`
theme `khronos.css` was based on, as well as Khronos-specific changes
for better accessibility and for special treatment of Valid Usage
blocks.